Update patch set 1
Patch Set 1:

It stalled boot when encrypted. I removed it when trying to fix bootup with encryption. At least on otus, Bluetooth is working fine without the script, with and without encryption. It made sense to me as I saw that the script didn't really do anything on these devices.

Patch-set: 1